Combine and reorganize fixture modules
